+++ lib/train/rank_range.dart
@@ -1,5 +1,7 @@
+import 'package:flutter/foundation.dart';
 import 'package:wqhub/wq/rank.dart';
 
+@immutable
 class RankRange {
   final Rank from;
   final Rank to;
@@ -13,6 +15,17 @@ class RankRange {
   bool isSingle() => from == to;
 
   bool isInside(RankRange that) => that.from <= from && to <= that.to;
+
+  @override
+  int get hashCode => Object.hash(from, to);
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    if (other.runtimeType != runtimeType) return false;
+
+    return other is RankRange && other.from == from && other.to == to;
+  }
 
   @override
   String toString() =>